+++ packages/eufemia/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.0
+
+**Breaking changes**
+
+- Update example
+- Bump `eufemia_components` to 2.0.0:
+
+> Rename button namespace to avoid conflict with new Material buttons:
+>
+> - `PrimaryButton` → `EufemiaPrimaryButton`
+> - `SecondaryButton` → `EufemiaSecondaryButton`
+> - `TextButton` → `EufemiaTextButton`
+> - `VividButton` → `EufemiaVividButton`
+
 ## 1.0.5+1
 
 - Update readme